refactor: correct code
