+++ src/app/shared/models/station.model.ts
@@ -1,0 +1,12 @@
+export interface IStationItem {
+  id: number;
+  city: string;
+  latitude: number;
+  longitude: number;
+  connectedTo: IConnectedStation[];
+}
+
+interface IConnectedStation {
+  id: number;
+  distance: number;
+}
